requirements.txt
flask>=2.0.0
openai>=1.0.0
requests>=2.0.0
python-dotenv>=0.19.0
Pillow>=9.0.0
werkzeug>=2.0.0
